Extract orphan reconciliation to satisfy source-file-length gate (issue #332)
Slice 3 inlined orphan_evidence + reap_orphaned_agent and their tests into
app_init.rs, pushing it to 1024 lines and tripping the CI hard limit (1000).
Extract those helpers and the three orphan-classification tests into a new
app_init_orphan_reconcile.rs sibling module (mirroring the existing
app_init_shell_reconcile.rs / app_init_process_binding.rs split).

Behavior is unchanged; the 24 app_init tests and 754 bin tests still pass.
app_init.rs is back under the limit; the new module is 124 lines.
